requirements.txt
attrs==21.4.0
click==8.0.3
coverage==6.3.1
Flask==2.0.2
freezegun==1.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==3.7.0
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
tomli==2.0.0
Werkzeug==2.0.2